third_party/WebKit/LayoutTests/media/mediasession/mojo/set-null-metadata.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

async_test(function(t) {
  mediaSessionServiceMock.then(m => {
    m.setMetadataCallback(t.step_func(function(receivedMetadata) {
      assert_equals(receivedMetadata, null);
      t.done();
    }));
    window.navigator.mediaSession.metadata = null;
  });
}, "test that null MediaMetadata is correctly propagated");

</script>
